refactor: order service 폴더구조 및 repository이름 변경
