+++ src/main/java/org/pokeherb/orderservice/application/service/OrderGetService.java
@@ -2,7 +2,7 @@ package org.pokeherb.orderservice.application.service;
 
 import lombok.RequiredArgsConstructor;
 import org.pokeherb.orderservice.domain.entity.Order;
-import org.pokeherb.orderservice.domain.OrderRepository;
+import org.pokeherb.orderservice.domain.repository.OrderRepository;
 import org.pokeherb.orderservice.domain.exception.OrderErrorCode;
 import org.pokeherb.orderservice.global.infrastructure.exception.CustomException;
 import org.pokeherb.orderservice.application.service.dto.response.OrderResponse;
