+++ src/main/java/org/pokeherb/orderservice/application/service/OrderSearchService.java
@@ -2,7 +2,7 @@ package org.pokeherb.orderservice.application.service;
 
 import lombok.RequiredArgsConstructor;
 import org.pokeherb.orderservice.application.service.dto.request.OrderSearchConditionRequestDto;
-import org.pokeherb.orderservice.domain.OrderQueryDao;
+import org.pokeherb.orderservice.domain.repository.OrderQueryRepository;
 import org.pokeherb.orderservice.application.service.dto.response.OrderSummaryResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -13,7 +13,7 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class OrderSearchService {
 
-    private final OrderQueryDao orderQueryDao;
+    private final OrderQueryRepository orderQueryDao;
 
     @Transactional
     public Page<OrderSummaryResponse> searchOrders(OrderSearchConditionRequestDto condition, Pageable pageable) {
